+++ packages/terra-form-textarea/src/TextareaField.jsx
@@ -76,10 +76,6 @@ const propTypes = {
    */
   onChange: PropTypes.func,
   /**
-   * Placeholder text.
-   */
-  placeholder: PropTypes.string,
-  /**
    * Whether or not the field is required.
    */
   required: PropTypes.bool,
@@ -111,7 +107,6 @@ const defaultProps = {
   labelAttrs: {},
   maxWidth: undefined,
   onChange: undefined,
-  placeholder: undefined,
   required: false,
   showOptional: false,
   value: undefined,
@@ -137,7 +132,6 @@ const TextareaField = (props) => {
     required,
     showOptional,
     onChange,
-    placeholder,
     value,
     ...customProps
   } = props;
@@ -179,7 +173,6 @@ const TextareaField = (props) => {
         id={inputId}
         isIncomplete={isIncomplete}
         onChange={onChange}
-        placeholder={placeholder || inputAttrs.placeholder}
         value={value}
         defaultValue={defaultValue}
         aria-describedby={ariaDescriptionIds}